requirements.txt
bcrypt==3.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
eventlet==0.26.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
google-api-core==1.22.1
google-auth==1.20.1
google-cloud-core==1.4.1
google-cloud-translate==2.0.1
googleapis-common-protos==1.52.0
greenlet==0.4.16
grpcio==1.31.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
protobuf==3.13.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-engineio==3.13.1
python-socketio==4.6.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
Werkzeug==1.0.1
WTForms==2.3.3
zope.event==4.4
zope.interface==5.1.0
